pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "midt"
version = "0.1.0"
description = "Mechanistic Interpretability of Decision Transformers"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Sebastian Lee"}
]

dependencies = [
    # Core
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "gymnasium>=0.28.0",

    # RL
    "stable-baselines3>=2.0.0",

    # Data
    "h5py>=3.8.0",
    "pandas>=2.0.0",

    # Config
    "pyyaml>=6.0",
    "pydantic>=2.0.0",

    # Logging
    "wandb>=0.15.0",
    "tensorboard>=2.13.0",
    "tqdm>=4.65.0",

    # Visualization
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.0.270",
    "mypy>=1.3.0",
    "ipykernel>=6.23.0",
    "jupyter>=1.0.0",
]

[project.scripts]
train-dqn = "midt.scripts.train_dqn:main"
train-dt = "midt.scripts.train_dt:main"
evaluate-dt = "midt.scripts.evaluate:main"
run-interp = "midt.scripts.run_interp:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
midt = ["py.typed"]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
